Add a placeholdery index page.
In order to make sure things look right everywhere without having to
deal with fontstuffs, I used an SVG for the "Na" text. Since there
doesn't seem to be support for SVG font/glyph stuff in current browsers,
I've hackily made the text selectable by overlaying it with an invisible
copy of the same.

The centering trick is the same one used at kittens.sodium.online.